+++ client/starwhale/api/job.py
@@ -1,3 +1,3 @@
-from ._impl.job import step
+from ._impl.job import step, Context
 
-__all__ = ["step"]
+__all__ = ["step", "Context"]
