+++ client/starwhale/consts/env.py
@@ -4,8 +4,6 @@ class SWEnv:
     eval_version = "SW_EVALUATION_VERSION"
     status_dir = "SW_TASK_STATUS_DIR"
     log_dir = "SW_TASK_LOG_DIR"
-    result_dir = "SW_TASK_RESULT_DIR"
-    input_config = "SW_TASK_INPUT_CONFIG"
     dataset_uri = "SW_DATASET_URI"
     dataset_row_start = "SW_DATASET_ROW_START"
     dataset_row_end = "SW_DATASET_ROW_END"
