+++ client/starwhale/core/model/model.py
@@ -30,12 +30,12 @@ from starwhale.base.cloud import CloudRequestMixed, CloudBundleModelMixin
 from starwhale.utils.http import ignore_error
 from starwhale.base.bundle import BaseBundle, LocalStorageBundleMixin
 from starwhale.utils.error import NoSupportError, FileFormatError
-from starwhale.core.job.model import Parser
 from starwhale.utils.progress import run_with_progress_bar
 from starwhale.base.bundle_copy import BundleCopy
 from starwhale.core.job.scheduler import Scheduler
 
 from .store import ModelStorage
+from ...api._impl.job import Parser
 
 
 class ModelRunConfig:
@@ -214,12 +214,9 @@ class StandaloneModel(Model, LocalStorageBundleMixin):
         task_index: int = 0,
         kw: t.Dict[str, t.Any] = {},
     ) -> None:
-        from starwhale.api._impl.model import _RunConfig
 
         if typ not in (EvalTaskType.ALL, EvalTaskType.SINGLE):
             raise NoSupportError(typ)
-
-        _RunConfig.set_env(kw)
 
         _module = StandaloneModel.get_pipeline_handler(
             workdir=src_dir, yaml_name=model_yaml_name
@@ -241,6 +238,7 @@ class StandaloneModel(Model, LocalStorageBundleMixin):
             src_dir=src_dir,
             dataset_uris=dataset_uris,
             steps=_jobs[job_name],
+            **kw,
         )
 
         if typ == EvalTaskType.ALL:
