+++ example/mnist/mnist/ppl.py
@@ -6,9 +6,9 @@ import torch
 from PIL import Image
 from torchvision import transforms
 
-from starwhale.api._impl.wrapper import BaseEvaluation
 from starwhale.api.model import PipelineHandler
 from starwhale.api.metric import multi_classification
+from starwhale.api.job import Context
 
 from .model import Net
 
@@ -18,8 +18,8 @@ ONE_IMAGE_SIZE = IMAGE_WIDTH * IMAGE_WIDTH
 
 
 class MNISTInference(PipelineHandler):
-    def __init__(self, evaluation: BaseEvaluation, device="cpu") -> None:
-        super().__init__(evaluation=evaluation, merge_label=True, ignore_error=True)
+    def __init__(self, context: Context, device="cpu") -> None:
+        super().__init__(context=context, merge_label=True, ignore_error=True)
         self.device = torch.device(device)
         self.model = self._load_model(self.device)
